requirements.txt
opencv-python-headless==4.7.0.72
open-clip-torch==2.0.0
torch==2.0.1
tqdm==4.67.1
matplotlib==3.7.1
pyyaml==6.0
numpy==1.23.0
pillow==10.2.0
